+++ workzone/source/uim/platform/workzone/application/usecases/manage/workpages.d
@@ -69,7 +69,12 @@ class ManageWorkpagesUseCase { // TODO: UIMUseCase {
     return CommandResult(true, page.id.value, "");
   }
 
-  void deleteWorkpage(TenantId tenantId, WorkpageId id) {
+  CommandResult deleteWorkpage(TenantId tenantId, WorkpageId id) {
+    auto entity = repo.findById(tenantId, id);
+    if (entity.isNull)
+      return CommandResult(false, "", "Page not found");
+
     repo.removeById(tenantId, id);
+    return CommandResult(true, entity.id.value, "");
   }
 }
